Add comment and create utils file
