+++ pkg/boa/utils.go
@@ -1,0 +1,25 @@
+package boa
+
+import (
+	"io"
+	"net/http"
+)
+
+func httpPostRequest(url string, body io.Reader, headers map[string]string) (*http.Response, error) {
+	req, err := http.NewRequest("POST", url, body)
+	if err != nil {
+		return nil, err
+	}
+
+	for k, v := range headers {
+		req.Header.Set(k, v)
+	}
+
+	client := &http.Client{}
+	resp, err := client.Do(req)
+	if err != nil {
+		return nil, err
+	}
+
+	return resp, nil
+}
